pyproject.toml
[project]
name = "aep-etl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "playwright>=1.55.0",
    "pydantic>=2.12.3",
    "pymongo>=4.15.3",
    "python-dotenv>=1.2.1",
    "tenacity>=9.1.2",
]
